Graphics and Sounds + Fixing Randomize Bug
fixed the randomize bug with the help of a script own RandomNumberGenerator

added some new graphics. background and new walls, yes they look pretty like the normal flappy bird but are still changable.

the game now has sounds and i added a sound when the player dies for later when the gamemanager handles the end of a game better

now the game starts with the press of a button and i changed how the game handles the start of a game